nearbyshops/requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.4.4
branca==0.4.1
certifi==2019.3.9
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudinary==1.24.0
defusedxml==0.6.0
Deprecated==1.2.11
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.39.1
django-countries==7.0
django-crispy-forms==1.10.0
django-crum==0.7.9
django-currentuser==0.5.2
django-debug-toolbar==3.2
django-ipware==3.0.2
django-localflavor==3.0.1
django-multiselectfield==0.1.12
django-phone-verify==2.0.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
#GDAL==3.1.4
gunicorn==20.0.4
geographiclib==1.50
geoip2==4.1.0
geopy==2.1.0
idna==2.8
idna-ssl==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.1.0
nexmo==2.5.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
phonenumbers==8.12.16
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.5.0
pycparser==2.20
#PyJWT==2.0.1
PyJWT==1.7.1
python-dotenv==0.15.0
python-stdnum==1.15
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
twilio==6.51.1
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
